refactor: add hook for hospitality item fetching
