+++ src/app/(site)/(apps-non-standard)/hospitality-hub/admin/components/CategoryTabContent.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import { SimpleGrid, Spinner } from "@chakra-ui/react";
+import useHospitalityItems from "../../hooks/useHospitalityItems";
 import HospitalityItemCard from "../../components/HospitalityItemCard";
 import {
   HospitalityCategory,
-  HospitalityItem,
 } from "../../hospitalityHubConfig";
 
 interface CategoryTabContentProps {
@@ -13,27 +12,7 @@ interface CategoryTabContentProps {
 }
 
 export const CategoryTabContent = ({ category }: CategoryTabContentProps) => {
-  const [items, setItems] = useState<HospitalityItem[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchItems = async () => {
-      setLoading(true);
-      try {
-        const res = await fetch(`/api/hospitality-hub/${category.key}`);
-        const data = await res.json();
-        if (res.ok) {
-          setItems(data.resource || []);
-        }
-      } catch (err) {
-        console.error(err);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchItems();
-  }, [category]);
+  const { items, loading } = useHospitalityItems(category.key);
 
   if (loading) {
     return <Spinner />;
